Add calculate method body and do its related tests
The body of calculate method is written, this codes check that the values for CDR and rates be valid by isValid methods in ChargingDetailRecord and ChargingDetailRecord classes.

Unit tests are done individually, a change is done on createChargingDetailRecord method in InvoiceTest to timestampStop create properly and a change is done on test_invoice_calculate_overall it was just a correct misspellings.