Add controller, its service and tests for making /rate API.
A service layer is made to proper data from request and call the calculate function from Invoice class with this parameter and send them as an array to controller. controller proper data and send them back.

A validation is added to validate the data that send to rate API.

ChargingDetailRecord is changed to calculate there value and duration in proper scale.

Unit invoice test is changed on its invoice_calculate_overall test to compare became true.

Tow test are added in feature invoice test. they check the returned response from API.

An api route is added to client can access /rate API.

A brief instruction for running is added to readme.